WIP move plugins folder out of src
