+++ plugins/pokedex-data-plugin/dex/details.js
@@ -1,4 +1,4 @@
-const { pokemonHeight, pokemonWeight } = require('../../../../__gamedata/data');
+const { pokemonHeight, pokemonWeight } = require('../../../__gamedata/data');
 
 const FEET_TO_CM = 30.48;
 const INCHES_TO_CM = 2.54;
