+++ plugins/pokedex-data-plugin/dex/details.test.js
