+++ plugins/pokedex-data-plugin/dex/functions.js
@@ -1,4 +1,4 @@
-const { PersonalTable } = require('../../../../__gamedata/data');
+const { PersonalTable } = require('../../../__gamedata/data');
 
 //BDSP does not stick to the same structure when working with forms, thus this map is necessary.
 const FORM_MAP = PersonalTable.Personal.reduce(createFormMap, {});
