+++ plugins/pokedex-data-plugin/dex/name.js
@@ -1,4 +1,4 @@
-const { PersonalTable, basePokemonNames, formPokemonNames } = require('../../../../__gamedata/data');
+const { PersonalTable, basePokemonNames, formPokemonNames } = require('../../../__gamedata/data');
 const { FORM_MAP } = require('./functions');
 
 const POKEMON_NAME_MAP = PersonalTable.Personal.reduce(createPokemonMap, {});
