+++ plugins/pokedex-data-plugin/dex/name.test.js
