+++ plugins/pokedex-data-plugin/dex/nature.test.js
