+++ plugins/pokedex-data-plugin/dex/pokemonFormSelector.js
@@ -1,4 +1,4 @@
-import { PersonalTable, formPokemonNames, basePokemonNames } from '../../../../__gamedata';
+import { PersonalTable, formPokemonNames, basePokemonNames } from '../../../__gamedata';
 
 const UNKNOWN_POKEMON_FORM_NAME = 'Egg';
