+++ plugins/pokedex-data-plugin/dex/types.js
@@ -1,4 +1,4 @@
-const { typeName } = require('../../../../__gamedata/data');
+const { typeName } = require('../../../__gamedata/data');
 
 function getTypeName(typeId = 0) {
   const type = typeName.labelDataArray[typeId]?.wordDataArray[0]?.str;
